Refactoring: Extract VerificationReport to its own file
